Implement player animations & Berta driving anims
Adds a script to stop the player and change it to the car
animations. It does some code stuff to stop the player at one point
where the transition is seamless.

This commit also moves the SpriteFrames animations to files in the
file system. That way we can re-use the animations more easily and
they can be referenced in code easier.